# api/openapi.yaml
openapi: 3.0.3
info:
  title: ShortWarden API
  version: 0.1.0
  description: |
    ShortWarden is a URL shortening service with user accounts, API keys for integrations,
    custom aliases, expiring links, click analytics, and domain management with DNS TXT verification.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/api-keys/{id}/domains:
    get:
      tags: [ApiKeys]
      summary: List allowed domains for API key
      description: Returns the list of domain IDs this API key is restricted to. Empty means unrestricted.
      operationId: listApiKeyDomains
      security:
        - CookieAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ApiKeys]
      summary: Replace allowed domains for API key
      description: Replaces the allowed domain IDs for this API key. Empty array means unrestricted.
      operationId: replaceApiKeyDomains
      security:
        - CookieAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uuid
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/links:
    get:
      tags: [Links]
      summary: List links
      description: Lists the authenticated user's links.
      operationId: listLinks
      security:
        - CookieAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: include_deleted
          in: query
          description: Include soft-deleted links.
          schema: { type: boolean, default: false }
        - name: tag
          in: query
          description: Optional tag filter (only links containing this tag).
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Link" }
    post:
      tags: [Links]
      summary: Create link
      description: Creates a short link with an optional custom alias and optional expiry.
      operationId: createLink
      security:
        - CookieAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /v1/stats:
    get:
      tags: [Users]
      summary: Get dashboard stats
      description: Returns high-level stats for the authenticated user.
      operationId: getStats
      security:
        - CookieAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
  /v1/home/recent-clicks:
    get:
      tags: [Users]
      summary: List recent clicks (dashboard)
      description: Returns recent click events across all links for the authenticated user.
      operationId: listRecentClicks
      security:
        - CookieAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RecentClick" }
  /v1/home/top-links:
    get:
      tags: [Users]
      summary: List top links (dashboard)
      description: Returns top links by clicks within a recent window for the authenticated user.
      operationId: listTopLinks
      security:
        - CookieAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 10 }
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 365, default: 7 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TopLink" }
...
  /v1/links/{id}:
    get:
      tags: [Links]
      summary: Get link
      description: Gets a link by id (must belong to the authenticated user).
      operationId: getLink
      security:
        - CookieAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Links]
      summary: Update link
      description: Updates a link's target, alias, expiry, tags, or domain.
      operationId: updateLink
      security:
        - CookieAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "409":
          $ref: "#/components/responses/Conflict"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Links]
      summary: Delete link
      description: Soft-deletes a link.
      operationId: deleteLink
      security:
        - CookieAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: sw_session
      description: HttpOnly session cookie set by login.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        API key authentication. Use: `Authorization: ApiKey <key>`.
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    Link:
      type: object
      required: [id, alias, target_url, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        domain_id: { type: string, format: uuid, nullable: true }
        alias: { type: string }
        target_url: { type: string, format: uri }
        title: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
        deleted_at: { type: string, format: date-time, nullable: true }
        click_count:
          type: integer
          format: int64
          description: Total click count for this link (all-time).
        short_url:
          type: string
          description: Convenience field with full short URL (domain + /r/alias).
    Stats:
      type: object
      required: [links_total, clicks_total, clicks_24h, clicks_7d]
      properties:
        links_total: { type: integer, format: int64 }
        clicks_total: { type: integer, format: int64 }
        clicks_24h: { type: integer, format: int64 }
        clicks_7d: { type: integer, format: int64 }
    RecentClick:
      type: object
      required: [id, clicked_at, link_id, alias]
      properties:
        id: { type: integer, format: int64 }
        clicked_at: { type: string, format: date-time }
        link_id: { type: string, format: uuid }
        alias: { type: string }
        domain_id: { type: string, format: uuid, nullable: true }
        referrer: { type: string, nullable: true }
        ip: { type: string, nullable: true }
        country: { type: string, nullable: true }
        device: { type: string, nullable: true }
        short_url:
          type: string
          nullable: true
          description: Convenience full short URL.
    TopLink:
      type: object
      required: [link, clicks]
      properties:
        link: { $ref: "#/components/schemas/Link" }
        clicks: { type: integer, format: int64 }
    CreateLinkRequest:
      type: object
      required: [target_url]
      properties:
        domain_id: { type: string, format: uuid, nullable: true }
        alias:
          type: string
          nullable: true
          description: Optional custom alias.
        target_url: { type: string, format: uri }
        title: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        expires_at: { type: string, format: date-time, nullable: true }
    UpdateLinkRequest:
      type: object
      required: [target_url]
      properties:
        domain_id: { type: string, format: uuid, nullable: true }
        alias:
          type: string
          nullable: true
          description: Optional new alias.
        target_url: { type: string, format: uri }
        title: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        expires_at: { type: string, format: date-time, nullable: true }
...